backend/pyproject.toml
[project]
name = "motifold"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.110.0",
    "sqlalchemy>=2.0.28",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "redis>=5.0.3",
    "celery>=5.3.6",
    "passlib[argon2]>=1.7.4",
    "PyJWT>=2.8.0",
    "pydantic-settings>=2.2.1",
    "sse-starlette>=2.0.0",
    "python-multipart>=0.0.9",
    "psycopg2-binary>=2.9.9",
    "openai>=1.14.0",
    "pgvector>=0.3.0",
    "psycopg-binary>=3.3.3",
    "socksio>=1.0.0",
    "pytest-mock>=3.15.1",
    "fastmcp>=3.2.0",
    "ddgs>=1.0.0",
    "aiohttp>=3.13.5",
    "beautifulsoup4>=4.14.3",
    "langgraph>=0.0.30",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langchain>=1.2.14",
    "langchain-openai>=0.1.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "watchfiles>=0.21.0",
]
ml = [
    "sentence-transformers>=3.0.0",
    "torch",
    "scipy",
]

[project.scripts]
motifold-local-mcp = "mcp_local:main"
